refactor #70: 드롭다운 열릴 방향을 결정하는 로직을 커스텀훅 분리
