+++ src/components/dropdown/Dropdown.tsx
@@ -1,7 +1,7 @@
-import { MouseEvent, ReactNode, useEffect, useRef, useState } from 'react';
+import { MouseEvent, ReactNode, useEffect, useState } from 'react';
 import { css, styled } from 'styled-components';
+import { useGetDropdownDirection } from '../../hooks/useGetDropdownDirection';
 import { useScrollLock } from '../../hooks/useScrollLock';
-import { useScreenConfigStore } from '../../stores/useScreenConfigStore';
 import { Button } from '../button/Button';
 import { Icon } from '../icon/Icon';
 import { IconsType } from '../icon/icons';
@@ -20,10 +20,8 @@ export function Dropdown({
   align = 'left',
 }: DropdownProps) {
   const [isOpen, setIsOpen] = useState(false);
-  const [direction, setDirection] = useState<'up' | 'down'>('down');
   const [lockScroll, unlockScroll] = useScrollLock('home--body__items');
-  const dropdownRef = useRef<HTMLDivElement>(null);
-  const screenRect = useScreenConfigStore(state => state.screenRect);
+  const { dropdownRef, direction, getDirection } = useGetDropdownDirection();
 
   useEffect(() => {
     if (isOpen) {
@@ -37,7 +35,7 @@ export function Dropdown({
   const onToggle = (event: MouseEvent) => {
     event.stopPropagation();
     if (!isOpen) {
-      chooseDirection();
+      getDirection();
     }
     setIsOpen(!isOpen);
   };
@@ -46,18 +44,6 @@ export function Dropdown({
     event.stopPropagation();
     unlockScroll();
     setIsOpen(false);
-  };
-
-  const chooseDirection = () => {
-    if (dropdownRef.current && screenRect) {
-      const { bottom } = dropdownRef.current.getBoundingClientRect();
-
-      if (screenRect.bottom - 64 < bottom + 230) {
-        setDirection('up');
-        return;
-      }
-      setDirection('down');
-    }
   };
 
   return (
